+++ contracts/OutcomeOracle.sol
@@ -3,14 +3,14 @@
 pragma solidity ^0.8.12;
 
 contract OutcomeOracle {
-    uint public numOutcomes;
+    uint256 public numOutcomes;
     address[] public arbiters;
-    mapping(address => uint) public votes;
+    mapping(address => uint256) public votes;
 
     string[] public outcomes;
 
     bool public decided;
-    uint public decision;
+    uint256 public decision;
 
     constructor(address[] memory arbiters_, string[] memory outcomes_) {
         arbiters = arbiters_;
@@ -18,23 +18,23 @@ contract OutcomeOracle {
         numOutcomes = outcomes.length;
     }
 
-    function vote(uint outcome) public {
+    function vote(uint256 outcome) public {
         votes[msg.sender] = outcome + 1;
     }
 
     function recordWinner() public {
         require(!decided, "No changing existing decision.");
 
-        uint[] memory voteCounts = new uint[](numOutcomes);
+        uint256[] memory voteCounts = new uint256[](numOutcomes);
 
-        for (uint i = 0; i < arbiters.length; i++) {
-            uint arbiterVote = votes[arbiters[i]] - 1;
+        for (uint256 i = 0; i < arbiters.length; i++) {
+            uint256 arbiterVote = votes[arbiters[i]] - 1;
             if (arbiterVote != 0) {
                 voteCounts[arbiterVote]++;
             }
         }
 
-        for (uint i = 0; i < numOutcomes; i++) {
+        for (uint256 i = 0; i < numOutcomes; i++) {
             if (3 * voteCounts[i] >= 2 * arbiters.length) {
                 // Supermajority vote.
                 (decided, decision) = (true, i);
